feat: scene renderer defined. Support classes are defined for using external storage and resources. Developing of raw meshes loader is in progress

+++ src/main/kotlin/com/conceptic/firefly/app/gl/support/Vector3.kt
@@ -1,27 +1,27 @@
 package com.conceptic.firefly.app.gl.support
 
-import org.joml.Vector3f
+import org.joml.Vector3d
 
-class Vector3(xCoord: Float, yCoord: Float, zCoord: Float) {
+class Vector3(xCoord: Double, yCoord: Double, zCoord: Double) {
     val x get() = components.x
     val y get() = components.y
     val z get() = components.z
     val normalized get() = components.normalize()
 
-    private var components = Vector3f(xCoord, yCoord, zCoord)
+    private var components = Vector3d(xCoord, yCoord, zCoord)
 
     fun distance(vector: Vector3) = components.distance(vector.x, vector.y, vector.z)
 
-    fun rotateX(angle: Float) = components.rotateX(angle)
+    fun rotateX(angle: Double) = components.rotateX(angle)
 
-    fun rotateY(angle: Float) = components.rotateY(angle)
+    fun rotateY(angle: Double) = components.rotateY(angle)
 
-    fun rotateZ(angle: Float) = components.rotateZ(angle)
+    fun rotateZ(angle: Double) = components.rotateZ(angle)
 
     fun cross(vector: Vector3) = components.cross(vector.x, vector.y, vector.z)
 
     companion object {
-        val ZERO = Vector3(0f, 0f, 0f)
+        val ZERO = Vector3(0.0, 0.0, 0.0)
     }
 
     override fun equals(other: Any?): Boolean {
